+++ packages/api/tests/bsky-agent.test.ts
@@ -236,6 +236,9 @@ describe('agent', () => {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.setAdultContentEnabled(true)
@@ -257,6 +260,9 @@ describe('agent', () => {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.setAdultContentEnabled(false)
@@ -277,6 +283,9 @@ describe('agent', () => {
         threadViewPrefs: {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
+        },
+        interests: {
+          tags: [],
         },
       })
 
@@ -300,6 +309,9 @@ describe('agent', () => {
         threadViewPrefs: {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
+        },
+        interests: {
+          tags: [],
         },
       })
 
@@ -325,6 +337,9 @@ describe('agent', () => {
         threadViewPrefs: {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
+        },
+        interests: {
+          tags: [],
         },
       })
 
@@ -353,6 +368,9 @@ describe('agent', () => {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.addPinnedFeed('at://bob.com/app.bsky.feed.generator/fake')
@@ -379,6 +397,9 @@ describe('agent', () => {
         threadViewPrefs: {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
+        },
+        interests: {
+          tags: [],
         },
       })
 
@@ -407,6 +428,9 @@ describe('agent', () => {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.removeSavedFeed('at://bob.com/app.bsky.feed.generator/fake')
@@ -434,6 +458,9 @@ describe('agent', () => {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.addPinnedFeed('at://bob.com/app.bsky.feed.generator/fake')
@@ -460,6 +487,9 @@ describe('agent', () => {
         threadViewPrefs: {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
+        },
+        interests: {
+          tags: [],
         },
       })
 
@@ -494,6 +524,9 @@ describe('agent', () => {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.removeSavedFeed('at://bob.com/app.bsky.feed.generator/fake')
@@ -520,6 +553,9 @@ describe('agent', () => {
         threadViewPrefs: {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
+        },
+        interests: {
+          tags: [],
         },
       })
 
@@ -548,6 +584,9 @@ describe('agent', () => {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.setFeedViewPrefs('home', { hideReplies: true })
@@ -575,6 +614,9 @@ describe('agent', () => {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.setFeedViewPrefs('home', { hideReplies: false })
@@ -601,6 +643,9 @@ describe('agent', () => {
         threadViewPrefs: {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
+        },
+        interests: {
+          tags: [],
         },
       })
 
@@ -636,6 +681,9 @@ describe('agent', () => {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.setThreadViewPrefs({ sort: 'random' })
@@ -670,6 +718,9 @@ describe('agent', () => {
           sort: 'random',
           prioritizeFollowedUsers: true,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.setThreadViewPrefs({ sort: 'oldest' })
@@ -703,6 +754,46 @@ describe('agent', () => {
         threadViewPrefs: {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
+        },
+        interests: {
+          tags: [],
+        },
+      })
+
+      await agent.setInterestsPref({ tags: ['foo', 'bar'] })
+      await expect(agent.getPreferences()).resolves.toStrictEqual({
+        feeds: {
+          pinned: ['at://bob.com/app.bsky.feed.generator/fake2'],
+          saved: ['at://bob.com/app.bsky.feed.generator/fake2'],
+        },
+        adultContentEnabled: false,
+        contentLabels: {
+          impersonation: 'hide',
+          spam: 'ignore',
+        },
+        birthDate: new Date('2023-09-11T18:05:42.556Z'),
+        feedViewPrefs: {
+          home: {
+            hideReplies: false,
+            hideRepliesByUnfollowed: false,
+            hideRepliesByLikeCount: 0,
+            hideReposts: false,
+            hideQuotePosts: false,
+          },
+          other: {
+            hideReplies: true,
+            hideRepliesByUnfollowed: false,
+            hideRepliesByLikeCount: 0,
+            hideReposts: false,
+            hideQuotePosts: false,
+          },
+        },
+        threadViewPrefs: {
+          sort: 'oldest',
+          prioritizeFollowedUsers: true,
+        },
+        interests: {
+          tags: ['foo', 'bar'],
         },
       })
     })
@@ -827,6 +918,9 @@ describe('agent', () => {
           sort: 'newest',
           prioritizeFollowedUsers: false,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.setAdultContentEnabled(false)
@@ -852,6 +946,9 @@ describe('agent', () => {
         threadViewPrefs: {
           sort: 'newest',
           prioritizeFollowedUsers: false,
+        },
+        interests: {
+          tags: [],
         },
       })
 
@@ -879,6 +976,9 @@ describe('agent', () => {
           sort: 'newest',
           prioritizeFollowedUsers: false,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.addPinnedFeed('at://bob.com/app.bsky.feed.generator/fake')
@@ -905,6 +1005,9 @@ describe('agent', () => {
           sort: 'newest',
           prioritizeFollowedUsers: false,
         },
+        interests: {
+          tags: [],
+        },
       })
 
       await agent.setPersonalDetails({ birthDate: '2023-09-11T18:05:42.556Z' })
@@ -930,6 +1033,9 @@ describe('agent', () => {
         threadViewPrefs: {
           sort: 'newest',
           prioritizeFollowedUsers: false,
+        },
+        interests: {
+          tags: [],
         },
       })
 
@@ -967,6 +1073,9 @@ describe('agent', () => {
         threadViewPrefs: {
           sort: 'oldest',
           prioritizeFollowedUsers: true,
+        },
+        interests: {
+          tags: [],
         },
       })