+++ packages/bsky/src/api/app/bsky/feed/getFeed.ts
@@ -17,7 +17,6 @@ import { OutputSchema as SkeletonOutput } from '../../../../lexicon/types/app/bs
 import { SkeletonFeedPost } from '../../../../lexicon/types/app/bsky/feed/defs'
 import { Server } from '../../../../lexicon'
 import AppContext from '../../../../context'
-import { AlgoResponse } from '../../../../feed-gen/types'
 import { Database } from '../../../../db'
 import {
   FeedHydrationState,
@@ -43,6 +42,7 @@ export default function (server: Server, ctx: AppContext) {
         authorization: req.headers['authorization'],
         'accept-language': req.headers['accept-language'],
       })
+      // @NOTE feed cursors should not be affected by appview swap
       const { timerSkele, timerHydr, resHeaders, ...result } = await getFeed(
         { ...params, viewer },
         {
@@ -70,16 +70,13 @@ const skeleton = async (
   ctx: Context,
 ): Promise<SkeletonState> => {
   const timerSkele = new ServerTimer('skele').start()
-  const localAlgo = ctx.appCtx.algos[params.feed]
   const feedParams: GetFeedParams = {
     feed: params.feed,
     limit: params.limit,
     cursor: params.cursor,
   }
   const { feedItems, cursor, resHeaders, ...passthrough } =
-    localAlgo !== undefined
-      ? await localAlgo(ctx.appCtx, params, params.viewer)
-      : await skeletonFromFeedGen(ctx, feedParams)
+    await skeletonFromFeedGen(ctx, feedParams)
   return {
     params,
     cursor,
@@ -150,6 +147,12 @@ type SkeletonState = {
 
 type HydrationState = SkeletonState &
   FeedHydrationState & { feedItems: FeedRow[]; timerHydr: ServerTimer }
+
+type AlgoResponse = {
+  feedItems: FeedRow[]
+  resHeaders?: Record<string, string>
+  cursor?: string
+}
 
 const skeletonFromFeedGen = async (
   ctx: Context,