+++ packages/dev-env/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @atproto/dev-env
 
+## 0.2.28
+
+### Patch Changes
+
+- Updated dependencies [[`8994d363`](https://github.com/bluesky-social/atproto/commit/8994d3633adad1c02569d6d44ae896e18195e8e2)]:
+  - @atproto/api@0.9.5
+  - @atproto/bsky@0.0.28
+  - @atproto/ozone@0.0.7
+  - @atproto/pds@0.3.16
+
+## 0.2.27
+
+### Patch Changes
+
+- Updated dependencies [[`4171c04a`](https://github.com/bluesky-social/atproto/commit/4171c04ad81c5734a4558bc41fa1c4f3a1aba18c)]:
+  - @atproto/api@0.9.4
+  - @atproto/bsky@0.0.27
+  - @atproto/ozone@0.0.6
+  - @atproto/pds@0.3.15
+
 ## 0.2.26
 
 ### Patch Changes